# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Subscription Platform API
  version: 0.7.1-stage7
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  description: >
    Stage 7 contract. Durable Telegram notification delivery and the bounded
    mTLS administrator API/CLI are executable locally. Production deployment
    and credential issuance remain out of scope.
servers:
  - url: https://api.example.invalid
    description: Public API placeholder
  - url: https://internal.example.invalid
    description: Internal API placeholder requiring mTLS service identity
  - url: https://admin.example.invalid
    description: Admin API placeholder requiring admin mTLS identity
  - url: https://sub.example.invalid
    description: Happ subscription endpoint placeholder
security:
  - ServiceMTLS: []
...
paths:
...
  /internal/v1/subscriptions/{subscription_id}/subscription-url/issue:
    post:
      tags: [Access]
      summary: Issue a subscription URL exactly once after access is ready
      operationId: issueSubscriptionUrl
      x-contract-status: implemented-stage5
      security:
        - ServiceMTLS: []
      x-allowed-service-identities: [telegram-bot]
      description: >
        Called synchronously by telegram-bot after the user presses "get VPN
        link". Access-service generates a new token, stores only its lookup
        HMAC-SHA-256, returns the URL once, and never persists plaintext. A replay of
        a completed idempotency key returns 409 because the secret response is
        intentionally not retained.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: New subscription URL issued once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionUrlIssued'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
  /internal/v1/subscriptions/{subscription_id}/subscription-url/rotate:
    post:
      tags: [Access]
      summary: Revoke old subscription URL token and issue a replacement
      operationId: rotateSubscriptionUrl
      x-contract-status: implemented-stage5
      security:
        - ServiceMTLS: []
      x-allowed-service-identities: [telegram-bot, admin-cli]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: Replacement subscription URL issued once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionUrlIssued'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
...
  /s/{token}:
    get:
      tags: [Access]
      summary: Return Happ-compatible subscription document
      operationId: getHappSubscription
      x-contract-status: implemented-stage5
      x-sensitive-path-parameter: token
      security:
        - SubscriptionUrlBearer: []
      description: >
        The bearer credential is carried in the path for Happ compatibility.
        Logs, metrics, traces, and reverse-proxy access logs must treat the path
        segment as a secret and record only the route template. Empty or nested
        malformed `/s/` paths receive the same generic no-store 404 response.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 43
            maxLength: 256
      responses:
        '200':
          description: Active subscription document
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            profile-title:
              schema:
                type: string
                maxLength: 25
            profile-update-interval:
              schema:
                type: string
                pattern: '^[0-9]+$'
            subscription-userinfo:
              schema:
                type: string
            support-url:
              schema:
                type: string
          content:
            text/plain:
              schema:
                type: string
                description: Happ-compatible lines and VLESS URIs. Must not be logged.
        '404':
          description: Same response for malformed paths and unknown, expired, or revoked tokens
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            text/plain:
              schema:
                type: string
                description: Generic unavailable response without identifiers.
        '429':
          description: Ephemeral public rate limit exceeded
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Retry-After:
              schema:
                type: string
        '503':
          description: Temporary internal failure with no credential details
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    ServiceMTLS:
      type: mutualTLS
      description: >
        Internal service identity from a trusted platform CA. Authorization is
        based on a strict SPIFFE URI from the verified certificate chain and
        per-endpoint allowed identities.
    ProvisioningMTLS:
      type: mutualTLS
      description: >
        Internal mTLS identity restricted to provisioning-service for fetching
        sensitive credential material from access-service.
    NodeAgentMTLS:
      type: mutualTLS
      description: >
        Private management-plane mTLS. The client must be provisioning-service,
        and provisioning-service additionally verifies the exact registered
        node SPIFFE URI in the server's verified certificate chain.
    AdminMTLS:
      type: mutualTLS
      description: >
        Admin CLI/client mTLS identity mapped to RBAC roles and audited actions.
    TelegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Telegram webhook secret token configured through Bot API setWebhook.
    SubscriptionUrlBearer:
      type: http
      scheme: bearer
      bearerFormat: URL path token
      description: >
        Bearer-secret semantics for `GET /s/{token}`. Happ v1 carries this
        token in the path, not in the Authorization header.
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    SubscriptionId:
      name: subscription_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Safe error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Idempotency conflict or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      additionalProperties: false
      properties:
        code:
          type: string
          maxLength: 128
        message:
          type: string
          maxLength: 512
        request_id:
          type: string
          maxLength: 128
        fields:
          type: array
          items:
            type: object
            required: [field, code]
            additionalProperties: false
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...
    SubscriptionUrlIssued:
      type: object
      required: [subscription_url]
      additionalProperties: false
      properties:
        subscription_url:
          type: string
          format: uri
          description: Returned only at issuance or rotation. Must not be logged or persisted by caller.
...